pyproject.toml
[project]
name = "tasman-pipeline-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-duckdb>=1.10.1",
    "dlt[duckdb,parquet]>=1.25.0",
    "marimo>=0.23.3",
    "pandas>=3.0.2",
    "requests>=2.33.1",
]
